requirements.txt
# 自动化汇总系统依赖包

# 数据处理
pandas>=1.3.0
openpyxl>=3.0.0

# 定时任务调度
schedule>=1.1.0

# 配置文件解析
PyYAML>=5.4.0

# 邮件发送
# 注意：Python标准库已包含smtplib和email模块

# HTTP请求（用于Webhook）
requests>=2.25.0

# 其他工具
python-dateutil>=2.8.0